Legacy SCCS files from Bivand (1992)
SYSTAT-COMPATIBLE SOFTWARE FOR MODELING SPATIAL DEPENDENCE AMONG OBSERVATIONS

Computers & Geosciences Vol. 18, No. 8, pp. 951-963, 1992

Abstract--Progress in making available SYSTAT-compatible software for modeling spatial dependence among observations is presented. The software--SPATIAL--is implemented on a range of UNIX workstations and under MS-DOS on Intel platforms. Estimation of simultaneous spatial regression models of a variety of classes is made possible, as is the reporting of simple spatial autocorrelation
statistics. The SYSTAT interface is constructed as a library of functions to be linked with module-specific analysis and reporting units using shared resources, permitting for example the graphical presentation of results and residuals within SYSTAT. The input of contiguity data for the spatial lattice under analysis from other applications is discussed in the context of interfaces between this and other software. 
....
The source code written by the author is available by anonymous ftp from barsoom.nhh.no subject to the usual restrictions on use for commercial purposes. It is packed both as a number of shar files, and as a compressed tar archive. On receipt of a diskette (3.5"
720 kbyte or 5.25" 1.2 Mbyte), the source code and MS-DOS executable will be sent by post. Other source code is accessible on "netlib" and "statlib" servers, while the GAL routines are available from their authors at the University of Newcastle (Dixon, Openshaw, and Wymer, 1988).